debian: Create systemd-preset package
The systemd-preset package centrally manages the
system services that are started based on "personality
types".

The systemd-preset package works in concert with the
platform-kickstart package. When the user selects the type of
node they want to install, the kickstart package
will symlink one of the node types from the systemd-preset
package and will run the systemctl preset-all command to
reset the services that are going to be started
at boot.

Test Plan
Build package
Build ISO
Install ISO
Bootstrap AIO install type

Story: 2009968
Task: 46406

Depends-On: https://review.opendev.org/c/starlingx/metal/+/854667
Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I0c9c78c7fb9a2e00904e934fde8b2ae9d7592380