Debian: Do not install gpu-operator into ISO
This change removes the gpu-operator helm chart from Debian ISO. The
NVIDIA GPU Operator uses the operator framework within Kubernetes to
automate the management of all NVIDIA software components needed to
provision GPU.
But if NVIDIA is not present in the nodes/controller this helm will
never be used and should not be in the ISO image.
The GPU Operator will continue to be built but not installed
into the image.

Test plan (Debian only)
PASS  build ISO and confirm that package was not installed

Story: 2009968
Task: 46344

Signed-off-by: Fabiano Mercer <fabiano.correamercer@windriver.com>
Change-Id: I1068eebd694cc9395dbe197257707abe2ff36e0e