Increase timeout for openstack commands
This commit adds a patch to the puppet-openstacklib module.

This puppet module is used by the bootstrap manifest to execute
openstack commands. After the change to debian, this 40s timeout
is not enough for some types of hardware.

Note: Further investigation is being done to understand why some
hardware is much slower on debian.

Test Plan:

PASS: Verify that installing a subcloud with redfish executes cleanly,
      the subcloud reaches online status and verify that the puppet
      bootstrap manifest runs without the openstack timeout error.

Closes-Bug: 1993585

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I3df79f2f6b7b5900808978914779cfe0c9949d72