Switch drivers in initramfs
Running depmod fails on Debian loads due to ostree. Using
"ostree admin unlock --hotfix" allows the command to pass,
but that's not usable for feature/product delivery.

So we will switch drivers in initramfs. If cmdline parameter
multi-drivers-switch is set to legacy, the legacy drivers
will be loaded by "insmod" command before udevd is started,
and that ensures that the legacy drivers will be loaded
rather than the latest drivers.

We need do the following steps to switch to the legacy drivers.
1. Add cmdline parameter multi-drivers-switch=cvl-2.54
2. reboot

Do the following steps to switch back to the latest drivers.
1. Add cmdline parameter multi-drivers-switch=cvl-4.0.1
2. reboot

Testing:
- An ISO image can be built successfully.
- PXE boot-based installation onto an All-in-One Duplex lab was
  successful with rt and std kernels.
- Latest driver versions are loaded by default and the DDP
  firmware version is correct. It is 1.3.30.0.
- It is possible to switch to the legacy drivers manually with
  command line parameter multi-drivers-switch=cvl-2.54, and the DDP
  firmware version is 1.3.26.0.
- It is also possible to switch back to the latest drivers manually
  with command line parameter multi-drivers-switch=cvl-4.0.1.

Closes-Bug: 1994914
Depends-On: https://review.opendev.org/c/starlingx/tools/+/862783

Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: I95922988e70615fe8efb1354f09c421972f014a0