CentOS: kdump: remove unnecessary bootargs
The 5.10.74 preempt-rt kernel reports the following warning when
dumping vmcore files due to the use of kernel command line arguments
such as nohz_full=, isolcpus=, rcu_nocbs= with the kexec/kdump kernel.

[    1.568059] WARNING: CPU: 0 PID: 0 at kernel/time/tick-sched.c:139
tick_sched_do_timer+0x5e/0x70
[    1.568064] Modules linked in:
[    1.568066] CPU: 0 PID: 0 Comm: swapper/0 Tainted: G          I
5.10.74-200.1648.tis.rt.el7.x86_64 #1
[    1.568068] Hardware name: Dell Inc. PowerEdge R740/0WRPXK, BIOS
2.10.2 02/24/2021
[    1.568068] RIP: 0010:tick_sched_do_timer+0x5e/0x70
[    1.568071] Code: 01 00 75 26 89 15 26 74 6f 01 48 8b 05 1b 87 d5 01

Commit 1655ee30e6("sched/isolation: really align nohz_full with
rcu_nocbs") is included in the 5.10.112 kernel, that had fixed the
warning. So the warning will not be reproduced with 5.10.112 and the
later versions of kernel.

We can remove the irqaffinity, isolcpus, nohz_full, rcu_nocbs, and
kthread_cpus arguments from the kdump kernel's command line arguments,
which will also fix the issue.

Testing:
- An ISO image can be built successfully.
- There are no warnings after the fix with 5.10.74 kernel.

Closes-Bug: 1997932

Signed-off-by: M. Vefa Bicakci <Vefa.Bicakci@windriver.com>
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Reported-by: M. Vefa Bicakci <Vefa.Bicakci@windriver.com>
Change-Id: I7d1dbd864fdfe2533197084d7274ef6ab70892db