kubelet CFS quota throttling for non integer cpulimit
A previous change set the cgroup cpu.cfs_quota_us value to -1 for
containers in pods in the Guaranteed QoS class.

We can only do this if we're allocating the entire CPU.  For non-
integer CPU allocations we need to set the cpu.cfs_quota_us value
to enforce the CPU limit configured on the container.

Test Plan:
Verified the pods that in the "Guaranteed" QoS class, on hosts that
have "kube-cpu-mgr-policy=static" have cpu.cfs_quota_us set to -1 for
integer cpu value.

Closes-Bug: 1997528

Signed-off-by: Boovan Rajendran <boovan.rajendran@windriver.com>
Change-Id: I33662e67706cee4cb0ce005bb09ce3b5fc717239