Remove unnecessary patch in pf-bb-config
The patch that checks for enable flag of the PCI device was re-added
back in the commit https://review.opendev.org/c/starlingx/integ/+/865422,
namely "Reject-device-configuration-if-not-enabled".
This patch is no longer necessary as it causes failure in configuring
the ACC200 device.
This commit removes that patch.

Test Plan:
Pass: Test host unlock with ACC100, ACC200 and N3000 devices.

Closes-Bug: 1999449

Signed-off-by: Teresa Ho <teresa.ho@windriver.com>
Change-Id: I5364c43d88d130948f55f2f60b1bb2b5f6f6ba77