Fix duplicate MTU value when running oam-modify
A system oam-modify command will apply the OAM addressing
changes at runtime. When running the oam-modify on a Debian system,
the OAM interface obtains the updated addressing,
but the state is "DOWN" after the runtime manifest is complete.

This results in a duplicate MTU value being generated by
the network_config puppet module inside each interface config file,
which prohibits ifup/ifdown of the interface.

Testing:

PASS: Using "system oam_modify" to verify that:
  - The OAM interface is "UP" with the proper addressing
  - The MTU value is not duplicated
  - A subsequent ifdown/ifup works properly

PASS: Ensure that the MTU value is configurable on all interfaces

Closes-Bug: 2003375

Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>
Change-Id: I16dfb81437d5503f34422372f5284dd894e4b311