Fix reconnection in logging.handlers.SysLogHandler
There are lots of exceptions reported in gunicorn.log.
That error was due to a bug present on SysLogHandler
that was reported on cpython repo but was not fixed for python3.9.2.

This change adds python3.9 package to the build system and patches
it in order to update SysLogHandler to fix the reconnection bug.

Test Plan:
PASS: Build python3.9 package. Install
libpython3.9-minimal_3.9.2-1.stx.1_amd64.deb and verify that the
exceptions are no longer present.

Depends-On: https://review.opendev.org/c/starlingx/root/+/873159
Closes-bug: 2006623

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: I6eb44544da5c05e712bc89e69193548667c8ab28