Add qemu-utils package to platform ISO
This change adds the qemu-utils package to StarlingX ISO.
This might be one utility of interest for users of our stx-openstack
application and was reported to be missing [1] when comparing to our
previous CentOS version.

After the Debian migration, QEMU was upversioned to be based on the
latest officially supported version for Debian Bullseye. The packages
are now split into new packages and qemu-utils is one of them.

To have the qemu-img binary available, qemu-utils needs to be installed
[2]. This binary allows one to create, convert and modify images
offline and handles all image formats supported by QEMU [3].

[1] bug/2007267
[2] https://salsa.debian.org/qemu-team/qemu/-/blob/debian/qemu_5.2+dfsg-11+deb11u2/debian/qemu-utils.install#L1
[3] https://www.qemu.org/docs/master/tools/qemu-img.html

TEST PLAN:
PASS - Ensure the qemu packages build generates qemu-utils*.deb
PASS - Ensure ISO can be built with qemu-utils*.deb installed
PASS - Bootstrap an AIO-SX system with the new ISO
PASS - Check the installed package and delivered files
       * apt list --installed | grep qemu-utils
       * dpkg -L qemu-utils
PASS - Ensure the "qemu-img" is found
       * qemu-img -V

Closes-Bug: 2007267

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: Ifcaab24540cfe6a922d1d08c4b63dc8a322cf437