Update debian package versions to use git commits
The Debian packaging has been changed to reflect all the
git commits under the directory, and not just the commits
to the metadata folder.

Package builder concatenates the number of revisions added since
the one defined in the BASE_SRCREV to the name of the .deb file
generated. Thus, any new submission to the code in these directories
increment this revision counter in the filename.

Test Plan:
PASS: Created new code submissions to package ceph and verify that
the version in the package name was incremented.
PASS: Created new code submissions to package parted and verify that
the version in the package name was incremented.
PASS: Created new code submissions to package trident-installer and
verify that the version in the package name was incremented.

Story: 2010550
Task: 47490

Signed-by: Pedro Vinícius Silva da Cruz <pedro.silvadacruz@windriver.com>
Change-Id: I16f834fd77f0abafaed39a0d4e6cd78d35fa4b98