chartmuseum: meta_data: remove explicit SRC_DIR
Chartmuseum package meta data file contains an explicit SRC_DIR with the
full path to itself. This prevents us from using this repo with
manifests where it is cloned in a different subdirectory, ie not under
${MY_REPO}/stx/integ/kubernetes/chartmuseum .

A recent commit to build-tools [1] allows us to omit this parameter as
it now defaults to project's top-level directory.

This patch removes SRC_DIR from meta_data.yaml, relying on build-pkg
being able to auto-detect it.

TESTS
=====
* Rebuild chartmuseum

[1] https://review.opendev.org/c/starlingx/root/+/876898

Story: 2010641
Task: 47621

Signed-off-by: Davlet Panech <davlet.panech@windriver.com>
Change-Id: I5e876f4971bcbac4b3cf60c95275298de79f6f88