helm: meta_data: remove explicit SRC_DIR
Helm package meta data file contains an explicit SRC_DIR with the
full path to itself. This prevents us from using this repo with
manifests where it is cloned in a different subdirectory, ie not under
${MY_REPO}/stx/integ/kubernetes/helm .

A recent commit to build-tools [1] allows us to omit this parameter as
it now defaults to project's top-level directory.

This patch removes SRC_DIR from meta_data.yaml, relying on build-pkg
being able to auto-detect it.

TESTS
=====
* Rebuild helm

[1] https://review.opendev.org/c/starlingx/root/+/876898

Story: 2010550
Task: 47622

Signed-off-by: Davlet Panech <davlet.panech@windriver.com>
Change-Id: Ie9f42ca84b03f1381f717b945e1d677d0b282076