installer: inc number of digits in scripts' names
The sections in kickstart cfg file are chopped up into scripts
named with number increased.
They are executed in the default sequence of 'ls' output, which
aren't aligned with the number in names when the number in name
is 10 and above.
To make the scripts run in sequence according to their names,
increase the number of digits in the names, so that 'ls' can list
them in number sequence.

Test plan:
 - PASS: Check about the installation log that "hook script"/"pre
   install script"/"post install script"/"post install nochroot script"
   run in right number sequence.
 - PASS: AIO-DX lib installation succeeds.

Closes-bug: 2012447

Signed-off-by: Li Zhou <li.zhou@windriver.com>
Change-Id: I80801d4c53dbb1867028f64d9bf51a774b8e12e7