Remove gpu-operator from build process
gpu-operator is currently not supported and can be removed from the
build process.

The package has issues and was removed from being supported in 2022. It
was supposed to be removed from the build accordingly. However, during
Helm upversioning it was found that the package was still being built
due to a linting error against the new Helm version.

Test plan:
PASS: build-pkgs -c -p gpu-operator
PASS: build-image

Story: 2010677
Task: 47810

Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>
Change-Id: I7229c5fb5fb78a708961527c1290147ddb9a6ed8