Remove broken/stale K8s upgrade script
upgrade_k8s_config.sh does not support
"flow" style of YAML.  This results in
the configmap becoming corrupted when
upgrading from k8s 1.21 to 1.22.

Removing the legacy code, the changes to the
config map are now handled in sysinv.

Test Plan
PASS:
- Manually update scripts on controllers and worker nodes based on
  https://review.opendev.org/c/starlingx/stx-puppet/+/880547?tab=comments
- Perform manual upgrade from k8s v1.21.8 to v1.22.5
- Verify kubernetes successfully upgraded to v1.22.5

Test was performed in the lab with local changes
to verify the code.
Patch was not tested.

Closes-Bug: 2017696
Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/880547
Signed-off-by: Gleb Aronsky <gleb.aronsky@windriver.com>
Change-Id: Ic47c3e40418898917934309f69b35ad1f401eb06