Update debian package revision counter
The Debian packaging has been changed to reflect all the
git commits under the directory, and not just the commits
to the metadata folder.

The SHA were chosen for being the migration to Debian of
the packages This ensures that any new code submissions under
those directories will increment the versions.

This commit won't change the versions of these specific 4
packages, so, for example: puppet-module-ceph version using
PKG_GITREVCOUNT was 6 and will remain 6 with GITREVCOUNT,
at least, until other code submission happens in the repo.

Test Plan:
 PASS: build-image
 PASS: build-pkgs -c -p puppet-module-ceph
 PASS: build-pkgs -c -p puppet-module-horizon
 PASS: build-pkgs -c -p puppet-module-keystone
 PASS: build-pkgs -c -p puppet-module-openstacklib
 PASS: manually make changes in the repo and check
       the version increasing

Story: 2010550
Task: 47859

Change-Id: I5cc6f55730bd9a7adc1027cab8171e3e2afb850d
Signed-off-by: Romulo Leite <romulo.leite@windriver.com>