Upgrade containerd to 1.6.21
As documented at
https://kubernetes.io/blog/2022/12/09/kubernetes-v1-26-release K8s
1.26 will no longer support containerd 1.5 and lower.

Accordingly we are upgrading containerd to 1.6.21.

The Debian package for containerd 1.6 comes from "bookworm" and
would require us to bring in a newer glibc and newer python3, which
is not something we want to do at this point.

As a workaround, we are pulling in the prebuilt binaries from
the containerd github project, which work fine with our existing
packages.

It's not a compile-time dependency, but these Ansible changes are
needed to actually run K8s 1.26:
https://review.opendev.org/c/starlingx/ansible-playbooks/+/879356

Also, the debian packaging tools do not allow binaries to be installed
to /usr/local/sbin, so the script that was there previously has
been moved to /usr/sbin.  There is a corresponding change in the
containerd service file override that should be merged at the same
time:

https://review.opendev.org/c/starlingx/config-files/+/883214

Test Plan:
PASS Install AIO-SX with K8s 1.25 and upgrade to K8s 1.26.
PASS Install with K8s 1.26 on AIO-SX and ensure the system comes up
     normally.

Story: 2010368
Task: 46590

Change-Id: I9e7a0e28a1b153141ff17a88efa0626314a01c91
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>