Adjust puppet-ceph dependency requirements
Puppet throws a few warnings during bootstrap because some
dependencies are not within the accepted version range. This
happens because those dependencies are at their latest
versions but we are running an older version of ceph in
order to have the ceph-disk functionality.
The proper fix would be upgrading ceph to a higher version,
but since there are plans in place to implement rook-ceph in
the future, this change increases the range of accepted
versions for the affected dependencies.

Test Plan:
PASS: Build & install
PASS: AIO-SX Successful Bootstrap
PASS: AIO-SX Successful Unlock
PASS: Verified that the 'unresolved dependencies' warnings for
      puppet-ceph are no longer present on puppet.log

Story: 2010757
Task: 48200

Change-Id: I4aec4d602650bfde026ce8894c08a1132bd79608
Signed-off-by: Matheus Guilhermino <matheus.machadoguilhermino@windriver.com>