Fix Ceph processes start race condition
When nodes are unlocked, mtc will start ceph processes after a
successful boot. It was also observed at the end of AIO-SX optimized
restore playbook.
This is causing racing condition with sm and pmon leading to failure
in some scenarios.

To avoid this, the script called by mtc will not start ceph processes
anymore. It will only set the flag to enable ceph to run on the node
and the processes will be started by sm or pmon later.

Test Plan:
  For each installation setup do:
    - Fresh install and verify Ceph is running with HEALTH_OK status;
    - Swact controllers and verify Ceph has HEALTH_OK status;
    - Run DOR (Dead Office Recover) and verify Ceph has HEALTH_OK
      status;
    - Lock/Unlock Controllers/Storage nodes and check Ceph has
      HEALTH_OK status;
    - Reboot active Controller and check Ceph has HEALTH_OK status.

  PASS: AIO-SX
  PASS: AIO-DX
  PASS: Standard (2+2)
  PASS: Standard with dedicated Storage (2+2+2)
  PASS: B&R AIO-SX
  PASS: B&R Optimized AIO-SX

Closes-bug: 2023445
Change-Id: I0c81749c6db1e17761aa8aca6276eff50f135959
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>