Present the correct version of ceph
The Ceph is reporting the wrong version when running the
commands:

ceph --version;
ceph mgr versions;
ceph tell osd.* version.

To avoid this, dl_hook was adjusted so that the ceph version
is returned and the SHA of the last commit is taken and saved
inside the .git_version file. Patch 0001 was removed as it
was not letting the .git_version file be read.

With this fix, the following output is shown after running
the 'ceph --version' command:

ceph version 14.2.22 (2ebd6bae80aca32269eb24d6471ebd72c22da03b)
nautilus (stable)

Test Plan:
PASS AIO-SX fresh install
PASS Run and check output ceph version commands

Closes-Bug: 2024681

Change-Id: I331e8b74b964b752e57b7a27b7c0c9054119ea51
Signed-off-by: Pedro Vinícius Silva da Cruz <pedro.silvadacruz@windriver.com>