Add the NFD namespace to the platform infrastructure list in kubelet
Actually, the pods installed in the node-feature-discovery namespace run
on application cores. Although, the Node Feature Discovery App is seen
as a platform app, and therefore, its pods need to run on platform
cores.

So, in this change, the node-feature-discovery namespace will be added
in the platform infrastructure list via kubelet patch.

Test Plan:
PASS - Verify if the pods are running at the platform cores

Story: 2010769
Task: 48327

Change-Id: If3fa8cb25050eb70c45bee54d7cf8fbfe9695d2d
Signed-off-by: Marcos Paulo Oliveira Silva <Marcos.PauloOliveiraSilva@windriver.com>