Remove openvswitch-switch-dpdk from stx-ovs image
During the migration of the `stx-ovs` container image to Debian (i.e. to
be based on Debian), a package was mistakenly included in the
Dockerfile: `openvswitch-switch-dpdk`.

The reason why this package is not needed is because, in our current
model, the use of OVS-DPDK only happens on the host side. On the
container side, there is no need to have a DPDK enabled implementation
of OVS.

One might wonder why this is only being removed now.

Well, since [1] was merged -- changing the `openvswitch-switch-dpdk`
build to consume static libraries instead of shared ones -- it has been
observed that stx-openstack is failing to apply *consistently*, due to
a problem in the `osh-openstack-openvswitch` helm release.

After reevaluating the packages necessary for the `stx-ovs` container
image build, it has been determined that the `openvswitch-switch-dpdk`
was mistakenly added there, as it does not suit our current model.

In fact, after removing it, the application has become applicable again.

[1] https://review.opendev.org/c/starlingx/integ/+/885663

Test Plan:
PASS - Build stx-ovs image
PASS - Manually upload built image to a system
PASS - Upload stx-openstack
PASS - Perform helm override to change container images
PASS - Apply stx-openstack
PASS - Ensure affected pods have successfully started and are running
PASS - Ensure affected pods liveness and readiness probes are healthy
PASS - Remove/delete stx-openstack

Closes-Bug: 2028061

Change-Id: I216d6ea06faebe4e31fda312ca72e401bce66240
Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>