Add sriov-fec-system namespace to the platform infra list in kubelet
Currently the pods installed in the sriov-fec-operator namespace run
on application cores. The sriov-fec-operator App is seen
as a platform app and therefore its pods need to run on platform
cores.

Accordingly, add the sriov-fec-system namespace to the list of
platform namespaces in kubelet.

Test Plan:
PASS - Verify if the pods are running at the platform cores

Story: 2010826
Task: 48628

Change-Id: I521deeef06a4073516c3a7d06ff8dd4308f7b6bb
Signed-off-by: Marcos Paulo Oliveira Silva <Marcos.PauloOliveiraSilva@windriver.com>