Fix the 'unless' condition of ceph-osd-prepare in osd puppet
In the 'unless' condition of ceph-prepare-osd-* there will be
a false positive if an exception occurs when running ceph-disk,
causing the osd to be formatted.

To fix this, the contents of the unless block were moved to the
command block and the execution of the binary (ceph-disk) was
isolated.

Test Plan:
PASS: Fresh install (AIO-SX)
PASS: Force exception in ceph-disk
PASS: Lock/Unlock the controller
PASS: Check if the osd has been wiped

Closes-bug: 2033552

Change-Id: I5374bc228eebabf4794e1ce302690dec258d6c2f
Signed-off-by: Erickson Silva <Erickson.SilvadeOliveira@windriver.com>