Select the clock matching requirements
Fix clock selection algorithm behavior where a clock source starts
to match requirements but is not selected because it has the same
ha_priority than active.

In a HA configuration with two or more clocks configured with equal
ha_priority if no clock source matches the requirements the first one
in the configuration file is selected active. This is a standard
behavior to always have a clock source, even when they are not
synchronized.

When one of the clock sources starts to match the requirements it
must be selected active, regardless of the priority.

But when a second clock source starts to match the requirements and
has the same ha_priority as the active, the active remains the clock
source. There is no need to switch active when they have equal
ha_priority.

Test plan: two sources with the same priority
PASS: Verify a clock source is selected active when it starts to match
the requirements and the current active doesn't match them, even if
they have equal ha_priority.
PASS: Verify a clock source isn't selected active when it starts to
match the requirements and the current active also matches them.

Regression: two sources with different priority
PASS: Verify a clock source is selected active when it starts to match
the requirements and the current active doesn't match them, even if
their ha_priority is lower than the actives.
PASS: Verify a clock source is selected active when it starts to match
the requirements and the current active also matches them but has
lower ha_priority configured.
PASS: Verify a clock source isn't selected active when it starts to
match the requirements and the current active also matches them
and has higher ha_priority configured.

Story: 2010723
Task: 48699

Change-Id: If80532b7b8febcc164f7c748a8d4122b4f10fd3b
Signed-off-by: Andre Mauricio Zelak <andre.zelak@windriver.com>