Fix timeout command in ceph-init-wrapper
When analyzing the ceph-process-states.log file, we observed a
recurring error scenario. In the /etc/init.d/ceph-init-wrapper
osd.0 script, the 'timeout' was consistently failing with error
code 125 on the execute_ceph_cmd function call. This failure was
due to the absence of a mandatory value parameter, causing
'timeout' to interpret 'ceph' as an invalid time interval.

To solve this bug, we introduced the necessary initialization
of the $WAIT_FOR_CMD variable. This ensures that the command is
executed correctly, addressing the issue and preventing the
recurrence of the 'timeout' error."

Test Plan:
  - PASS: Force the disk process to be reported as hung and
    check the aforementioned log for the desired output.

Closes-Bug: 2037728
Change-Id: Ic337b212b74c0cc76f25f4aaf9a99d77f8d9250d
Signed-off-by: Tiago Leal <Tiago.Leal@windriver.com>