kdump-tools: fix oom issue during kdump
The kdump initramfs has included NIC device drivers such as ice, iavf,
i40e, mlx5* and the like. Loading these drivers in a kdump environment
is problematic, because these drivers are known to consume a lot of
memory.

In addition, the systemd-sysctl.service systemd unit causes the
following sysctl to be set, which forces the kernel to keep more than
1 GiB of free memory, which results in out-of-memory conditions:
"vm.min_free_kbytes=1179648".

All of these introduce a possibility of not being able to collect a
vmcore file after a kernel crash. There is a need to exclude
unnecessary NIC drivers from the kdump environment and prevent the
sysctl settings from taking place to make vmcore collection more
reliable.

Verification:
- build-pkgs; build-iso; install and boot up on aio-sx lab.
- All these backlist drivers are not loaded in kdump kernel.

Closes-Bug: 2038804

Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: I820280b1674d09f42b6abfc25bfa07f44b4f7b44