Add puppet-memcached with FQDN support
The new release allows the use of hostname/FQDN to configure the
memcached. It is necessary to configure the parameter "listen"
instead of "listen_ip" (deprecated). The use of FQDN is necessary
to allow the reconfiguration of the MGMT network.

The management network is used extensively for all internal
communication. Since the original use of the network was a private
network before it was exposed for external communication in ai
distributed cloud configuration, it was never designed to be
reconfigured.

To support MGMT network reconfiguration the idea is to configure the
applications to use the hostname/FQDN instead of a static MGMT IP
address.

In this way, the MGMT network can be changed and the services and
applications will still work since they are using the hostname/FQDN
and the DNS will be responsible to translate to the current MGMT
IP address.

Given the complexities of supporting the multi-host reconfiguration,
this task will focus on support for AIO-SX only.

Test plan ( Debian only )
 - AIO-SX and AIO-DX virtualbox installation
 - netstat -tupl ( no services are using the MGMT IP address )

Story: 2010722
Task: 48464
Depends-on: https://review.opendev.org/c/starlingx/config/+/886208

Change-Id: Ic45b9ea8055d81dd778267bf77326ad4f641d166
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>