Fix uninitialized ts2phc variable in nmea_scan_rmc
This change pulls in an upstream linuxptp fix to initialize the tm_isdst
variable.

An unitialized tm_isdst variable in ts2phc can result in mktime failing
and cause ts2phc to be unable to sync time with a "invalid master time
stamp" error.

The fault was intermittent based on the random value in the unitialized
variable. If it was read as a positive integer, mktime would fail and
the symptom would occur.

The upstream commit id is:
63fc1ef4fd5e5fc45dd4de3bf27920bb109a4357

Test plan:
Pass: Verify package build
Pass: Deploy updated ts2phc binary and perform repeated service
start/stops. The fault was not reproduced after 20 attempts.

Closes-bug: https://bugs.launchpad.net/starlingx/+bug/2055464

Change-Id: I9fb1722c6ab93f6bb9ec6cdc4fbe902a823b3e2e
Signed-off-by: Cole Walker <cole.walker@windriver.com>