Switch drivers between OOT and in-tree
In order to select between the in-tree and OOT drivers, a cmdline
parameter multi-drivers-switch need be used. If cmdline parameter
multi-drivers-switch is set to cvl-4.0.1, the OOT drivers will be
loaded by "insmod" command before udevd is started, and that ensures
that the OOT drivers will be loaded rather than the in-tree drivers.

The default drivers will be the in-tree drivers if there is not
multi-drivers-switch in the cmdline parameters or it is not set
correct because the out-of-tree drivers have been moved to the the
"weak-updates" subdirectory, we would like them to have a lower
precedence/lower priority than the in-tree versions.

Note: The in-tree driver ice will use the OOT ice driver's ice.pkg
file, The version is 1.3.30.0. We will wait Intel's test result and
give feedback which DDP package should be used for the in-tree
drivers, and do the related update.

Verification:
Switch drivers between the OOT and in-tree drivers.
- switch to the OOT drivers
  1. Add cmdline parameter multi-drivers-switch=cvl-4.0.1
  2. reboot
  can switch to the OOT drivers.
- switch to the in-tree drivers
  1. Remove cmdline parameter multi-drivers-switch=cvl-4.0.1
  2. reboot
  can switch to the in-tree drivers.

Story: 2011056
Task: 49673

Depends-On: https://review.opendev.org/c/starlingx/tools/+/911401
Change-Id: I9a89d5a541b9da822fe5e079b919df3caf1acffc
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>